Separated logic of mapping csv rows into entities
